requirements.txt
alabaster==0.7.10
amqp==2.1.4
appdirs==1.4.3
appnope==0.1.0
argh==0.26.2
Babel==2.3.4
bcrypt==3.1.1
billiard==3.5.0.2
celery==4.0.2
certifi==2017.1.23
cffi==1.8.3
decorator==4.0.10
defusedxml==0.4.1
Django==1.10.3
django-bootstrap3==7.1.0
django-braces==1.11.0
django-celery-beat==1.0.1
django-cors-headers==2.1.0
django-el-pagination==3.0.1
django-filter==1.0.0
django-haystack==2.6.0
django-model-utils==3.0.0
django-semanticui-form==0.0.1
django-star-ratings==0.5.5
djangorestframework==3.5.1
djangorestframework-jwt==1.11.0
docutils==0.13.1
elasticsearch==5.1.0
flower==0.9.1
future==0.16.0
gunicorn==19.6.0
imagesize==0.7.1
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.9.6
kombu==4.0.2
livereload==2.5.1
MarkupSafe==1.0
mysqlclient==1.3.9
oauthlib==2.0.0
packaging==16.8
pathtools==0.1.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
port-for==0.3.1
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycparser==2.14
Pygments==2.1.3
PyJWT==1.5.2
PyMySQL==0.7.10
pyparsing==2.2.0
python-social==0.1
python-social-auth==0.2.21
python-twitter==3.2
python3-openid==3.0.10
pytz==2016.10
PyYAML==3.12
requests==2.12.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
social-auth-app-django==1.0.1
social-auth-core==1.1.0
tornado==4.2
traitlets==4.3.1
urllib3==1.20
vine==1.1.3
watchdog==0.8.3
wcwidth==0.1.7
